Add modal backdrop
Adds a dark, transparent, full-page backdrop behind the widget. It's
shown by default on small screens, but not on large ones. The config
option `modalBackdrop` can be set to `true` to always show it, `false`
to never show it, or `onlySmallScreens` (the default).

This also changes the small-screen modal to not use the entire height of
the screen, which felt a bit awkward to me in actual usage.